combine diagbostics for sycl kernel anfd function
