+++ clang/test/SemaSYCL/reqd-sub-group-size.cpp
@@ -1,4 +1,50 @@
-// RUN: %clang_cc1 -fsycl-is-device -verify %s
+// RUN: %clang_cc1 -fsycl-is-device -internal-isystem %S/Inputs -fsyntax-only -sycl-std=2017 -Wno-sycl-2017-compat -verify -pedantic %s
+
+// The test checks functionality of [[intel::reqd_sub_group_size()]] attribute on SYCL kernel.
+
+#include "sycl.hpp"
+
+using namespace cl::sycl;
+queue q;
+
+[[intel::reqd_sub_group_size(4)]] void foo() {} // expected-note {{conflicting attribute is here}}
+// expected-note@-1 {{conflicting attribute is here}}
+[[intel::reqd_sub_group_size(32)]] void baz() {} // expected-note {{conflicting attribute is here}}
+
+class Functor8 { // expected-error {{conflicting attributes applied to a SYCL kernel}}
+public:
+  [[intel::reqd_sub_group_size(8)]] void operator()() const { // expected-note {{conflicting attribute is here}}
+    foo();
+  }
+};
+
+int main() {
+  q.submit([&](handler &h) {
+    Functor8 f8;
+    h.single_task<class kernel_name1>(f8);
+
+    h.single_task<class kernel_name2>([]() { // expected-error {{conflicting attributes applied to a SYCL kernel}}
+      foo();
+      baz();
+    });
+  });
+  return 0;
+}
+[[intel::reqd_sub_group_size(16)]] SYCL_EXTERNAL void B();
+[[intel::reqd_sub_group_size(16)]] void A() {
+}
+
+[[intel::reqd_sub_group_size(16)]] SYCL_EXTERNAL void B() {
+  A();
+}
+// expected-note@+1 {{conflicting attribute is here}}
+[[intel::reqd_sub_group_size(2)]] void sg_size2() {}
+
+// expected-note@+2 {{conflicting attribute is here}}
+// expected-error@+1 {{conflicting attributes applied to a SYCL kernel}}
+[[intel::reqd_sub_group_size(4)]] __attribute__((sycl_device)) void sg_size4() {
+  sg_size2();
+}
 
 // Test that checks support and functionality of reqd_sub_group_size attribute support on function.
 
@@ -44,14 +90,14 @@ void test() {
 
 // Test that checks expression is not a constant expression.
 // expected-note@+1{{declared here}}
-int foo();
+int foo1();
 // expected-error@+2{{expression is not an integral constant expression}}
-// expected-note@+1{{non-constexpr function 'foo' cannot be used in a constant expression}}
-[[intel::reqd_sub_group_size(foo() + 12)]] void func1();
+// expected-note@+1{{non-constexpr function 'foo1' cannot be used in a constant expression}}
+[[intel::reqd_sub_group_size(foo1() + 12)]] void func1();
 
 // Test that checks expression is a constant expression.
-constexpr int bar() { return 0; }
-[[intel::reqd_sub_group_size(bar() + 12)]] void func2(); // OK
+constexpr int bar1() { return 0; }
+[[intel::reqd_sub_group_size(bar1() + 12)]] void func2(); // OK
 
 // Test that checks template parameter support on member function of class template.
 template <int SIZE>
